fix CMP instruction

+++ src/cpu.rs
@@ -571,7 +571,7 @@ impl<M: Memory, I: InterruptProvider> Cpu<M, I> {
     fn cmp(&mut self, a: u8, addressing_mode: AddressingMode) {
         let m = self.read_operand(addressing_mode);
         self.p.set_zero(a == m);
-        self.p.set_negative(a < m);
+        self.p.set_negative((a as i8) < (m as i8));
         self.p.set_carry(a >= m);
     }
